+++ ui/apps/dashboard/src/gql/gql.ts
@@ -82,6 +82,7 @@ const documents = {
     "\n  query GetAccountSupportInfo {\n    account {\n      id\n      plan {\n        id\n        name\n        amount\n        features\n      }\n    }\n  }\n": types.GetAccountSupportInfoDocument,
     "\n  query GetArchivedAppBannerData($envID: ID!, $externalAppID: String!) {\n    environment: workspace(id: $envID) {\n      app: appByExternalID(externalID: $externalAppID) {\n        isArchived\n      }\n    }\n  }\n": types.GetArchivedAppBannerDataDocument,
     "\n  query GetArchivedFuncBannerData($envID: ID!, $funcID: ID!) {\n    environment: workspace(id: $envID) {\n      function: workflow(id: $funcID) {\n        id\n        archivedAt\n      }\n    }\n  }\n": types.GetArchivedFuncBannerDataDocument,
+    "\n  query EntitlementUsage {\n    account {\n      id\n      entitlementUsage {\n        runCount {\n          current\n          limit\n        }\n      }\n    }\n  }\n": types.EntitlementUsageDocument,
     "\n  mutation ArchiveEnvironment($id: ID!) {\n    archiveEnvironment(id: $id) {\n      id\n    }\n  }\n": types.ArchiveEnvironmentDocument,
     "\n  mutation UnarchiveEnvironment($id: ID!) {\n    unarchiveEnvironment(id: $id) {\n      id\n    }\n  }\n": types.UnarchiveEnvironmentDocument,
     "\n  mutation DisableEnvironmentAutoArchiveDocument($id: ID!) {\n    disableEnvironmentAutoArchive(id: $id) {\n      id\n    }\n  }\n": types.DisableEnvironmentAutoArchiveDocumentDocument,
@@ -94,7 +95,6 @@ const documents = {
     "\n  query GetFunctionVersionNumber($slug: String!, $environmentID: ID!) {\n    workspace(id: $environmentID) {\n      workflow: workflowBySlug(slug: $slug) {\n        id\n        isPaused\n        name\n        archivedAt\n        current {\n          version\n        }\n        previous {\n          version\n        }\n      }\n    }\n  }\n": types.GetFunctionVersionNumberDocument,
     "\n  mutation PauseFunction($fnID: ID!, $cancelRunning: Boolean) {\n    pauseFunction(fnID: $fnID, cancelRunning: $cancelRunning) {\n      id\n    }\n  }\n": types.PauseFunctionDocument,
     "\n  mutation UnpauseFunction($fnID: ID!) {\n    unpauseFunction(fnID: $fnID) {\n      id\n    }\n  }\n": types.UnpauseFunctionDocument,
-    "\n  query EntitlementUsage {\n    account {\n      id\n      entitlementUsage {\n        runCount {\n          current\n          limit\n        }\n      }\n    }\n  }\n": types.EntitlementUsageDocument,
     "\n  query MetricsLookups($envSlug: String!, $page: Int, $pageSize: Int) {\n    envBySlug(slug: $envSlug) {\n      apps {\n        externalID\n        id\n        name\n        isArchived\n      }\n      workflows @paginated(perPage: $pageSize, page: $page) {\n        data {\n          name\n          id\n          slug\n        }\n        page {\n          page\n          totalPages\n          perPage\n        }\n      }\n    }\n  }\n": types.MetricsLookupsDocument,
     "\n  query FunctionStatusMetrics(\n    $workspaceId: ID!\n    $from: Time!\n    $functionIDs: [UUID!]\n    $appIDs: [UUID!]\n    $until: Time\n    $scope: MetricsScope!\n  ) {\n    workspace(id: $workspaceId) {\n      scheduled: scopedMetrics(\n        filter: {\n          name: \"function_run_scheduled_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      started: scopedMetrics(\n        filter: {\n          name: \"function_run_started_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      completed: scopedMetrics(\n        filter: {\n          name: \"function_run_ended_total\"\n          scope: $scope\n          groupBy: \"status\"\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      completedByFunction: scopedMetrics(\n        filter: {\n          name: \"function_run_ended_total\"\n          scope: FN\n          groupBy: \"status\"\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      totals: scopedFunctionStatus(\n        filter: {\n          name: \"function_run_scheduled_total\"\n          scope: FN\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        queued\n        running\n        completed\n        failed\n        cancelled\n        cancelled\n        skipped\n      }\n    }\n  }\n": types.FunctionStatusMetricsDocument,
     "\n  query VolumeMetrics(\n    $workspaceId: ID!\n    $from: Time!\n    $functionIDs: [UUID!]\n    $appIDs: [UUID!]\n    $until: Time\n    $scope: MetricsScope!\n  ) {\n    workspace(id: $workspaceId) {\n      runsThroughput: scopedMetrics(\n        filter: {\n          name: \"function_run_ended_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      sdkThroughputEnded: scopedMetrics(\n        filter: {\n          name: \"sdk_req_ended_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      sdkThroughputStarted: scopedMetrics(\n        filter: {\n          name: \"sdk_req_started_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      sdkThroughputScheduled: scopedMetrics(\n        filter: {\n          name: \"sdk_req_scheduled_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      stepThroughput: scopedMetrics(\n        filter: {\n          name: \"step_output_bytes_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      backlog: scopedMetrics(\n        filter: {\n          name: \"steps_scheduled\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      stepRunning: scopedMetrics(\n        filter: {\n          name: \"steps_running\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      concurrency: scopedMetrics(\n        filter: {\n          name: \"concurrency_limit_reached_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n  }\n": types.VolumeMetricsDocument,
@@ -421,6 +421,10 @@ export function graphql(source: "\n  query GetArchivedFuncBannerData($envID: ID!
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
+export function graphql(source: "\n  query EntitlementUsage {\n    account {\n      id\n      entitlementUsage {\n        runCount {\n          current\n          limit\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query EntitlementUsage {\n    account {\n      id\n      entitlementUsage {\n        runCount {\n          current\n          limit\n        }\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
 export function graphql(source: "\n  mutation ArchiveEnvironment($id: ID!) {\n    archiveEnvironment(id: $id) {\n      id\n    }\n  }\n"): (typeof documents)["\n  mutation ArchiveEnvironment($id: ID!) {\n    archiveEnvironment(id: $id) {\n      id\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
@@ -466,10 +470,6 @@ export function graphql(source: "\n  mutation PauseFunction($fnID: ID!, $cancelR
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(source: "\n  mutation UnpauseFunction($fnID: ID!) {\n    unpauseFunction(fnID: $fnID) {\n      id\n    }\n  }\n"): (typeof documents)["\n  mutation UnpauseFunction($fnID: ID!) {\n    unpauseFunction(fnID: $fnID) {\n      id\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query EntitlementUsage {\n    account {\n      id\n      entitlementUsage {\n        runCount {\n          current\n          limit\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query EntitlementUsage {\n    account {\n      id\n      entitlementUsage {\n        runCount {\n          current\n          limit\n        }\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
